Moved ListeningIndicator to the components directory
