+++ packages/cli/src/ui/components/InputPrompt.test.tsx
@@ -104,11 +104,9 @@ vi.mock('../hooks/useReverseSearchCompletion.js');
 vi.mock('clipboardy');
 vi.mock('../utils/clipboardUtils.js');
 vi.mock('../hooks/useKittyKeyboardProtocol.js');
-
-vi.mock('./shared/ListeningIndicator.js', () => ({
+vi.mock('./ListeningIndicator.js', () => ({
   ListeningIndicator: vi.fn(({ color }) => <Text color={color}>~~~ </Text>),
 }));
-
 // Mock ink BEFORE importing components that use it to intercept terminalCursorPosition
 vi.mock('ink', async (importOriginal) => {
   const actual = await importOriginal<typeof import('ink')>();
