+++ packages/cli/src/ui/components/ListeningIndicator.tsx
